+++ src/core/ParserTools/ParserLine.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import { CAhkInclude } from '../../AhkSymbol/CAhkInclude';
 import type { CAhkComment, CAhkLabel, TLineClass } from '../../AhkSymbol/CAhkLine';
 import { CAhkDirectives } from '../../AhkSymbol/CAhkLine';
+import { DirectivesMDMap } from '../../tools/Built-in/Directives.tool';
 import { getRangeOfLine } from '../../tools/range/getRangeOfLine';
 import type { TFuncInput } from '../getChildren';
 import { getComment } from './getComment';
@@ -49,18 +50,23 @@ export function ParserLine(FuncInput: TFuncInput): CAhkComment | TLineClass | nu
     /**
      * just of #NoEnv [#$@\w\u{A1}-\u{FFFF}]+
      */
-    const ma: RegExpMatchArray | null = strTrim.match(/^(#\w+)\b[ \t]*(?![:+\-*/.|&^]=)/u);
+    const ma: RegExpMatchArray | null = strTrim.match(/^(#\w+)[ \t]/u);
     if (ma === null) return null;
     const name: string = ma[1];
     const col: number = lStr.indexOf('#');
-    return new CAhkDirectives({
-        name,
-        range: rangeOfLine,
-        selectionRange: new vscode.Range(
-            new vscode.Position(line, col),
-            new vscode.Position(line, col + name.length),
-        ),
-        uri,
-        AhkTokenLine,
-    });
+
+    const hashtag: string = name.replace('#', '').toUpperCase();
+    if (DirectivesMDMap.has(hashtag)) {
+        return new CAhkDirectives({
+            name,
+            range: rangeOfLine,
+            selectionRange: new vscode.Range(
+                new vscode.Position(line, col),
+                new vscode.Position(line, col + name.length),
+            ),
+            uri,
+            AhkTokenLine,
+        }, hashtag);
+    }
+    return null;
 }